# PERF: restore performance for unsorted CategoricalDtype comparison